Fix the usage of lock_context_list and unlock_context_list.  perror was a mistake.
Add is_context_valid to validate a shared context.

Make the Apple context constructor return a bool.

Make the context constructor set an *errorptr to the X/GLX error that occurred, if needed. 

Remove some code that used gc->xid.  It was a potential problem, with an early
release/free.

Properly handle and report X errors in the glXCreateContext path.

Remove some dead code from various areas of glxcmds.c.  It was confusing matters.

Simplify glXWaitGL, and glXWaitX.

Fix the X error handling in the new glXCopyContext code.

Tested with Pymol, Coot, and Chimera.

The Togl stereo demo fails now in a different way.  It was triggering a libGL
abort.  Now it results in a BadMatch X error, perhaps because the shared context
is a valid context, but it's incompatible with the other context in terms of the 
PixelFormatObj it was created with.  This may be an incompatible difference between
the 2 OpenGL systems.



git-svn-id: http://svn.macosforge.org/repository/xquartz/AppleSGLX/trunk@102 9052f27e-8171-453d-b9bb-b0e8541e09c8