Make glXSelectEvent a no-op.
Perhaps in the future it should check if the GLXDrawable
is a valid GLXPbuffer, and save the mask for use with 
glXGetSelectedEvent.  

I'm going to look into adding that feature in a moment.

The glXSelectEvent shouldn't actually do anything, because
our Pbuffers can't be damaged, and we can't receive an 
event when they are, however it might be useful to have
the glXGetSelectedEvent match the input mask for the
glXSelectEvent.

There is also the issue of issuing an XError if the drawable
isn't a pbuffer or window.  Technically it makes no sense
to glXSelectEvent a window drawable, because there are no
event types for windows according to the spec.  So there
is no valid mask for a window.  I suspect this allowance 
was left in the spec for the future.



git-svn-id: http://svn.macosforge.org/repository/xquartz/AppleSGLX/trunk@187 9052f27e-8171-453d-b9bb-b0e8541e09c8