Update the glxext.h to the latest from the OpenGL registry.
I seems I was wrong to disable these type of patterns for
unsupported extensions:

#ifndef GLX_SOME_EXT
#define GLX_SOME_EXT 1
...
#endif

See: http://www.opengl.org/registry/ABI/#5

This brings in more features, and the build seems to work 
alright still.

It seems that the design is such that you're allowed to build
with all extensions, but the symbols may not actually exist,
so that's where glXGetProcAddress and friends come in.



git-svn-id: http://svn.macosforge.org/repository/xquartz/AppleSGLX/trunk@206 9052f27e-8171-453d-b9bb-b0e8541e09c8