Fix an issue with glDrawBuffer that may solve a Pymol problem.
The GL_BACK was being translated to a GL_BACK_LEFT and
GL_BACK_RIGHT correctly, however the GL_FRONT should also be
translated to GL_FRONT_LEFT and GL_FRONT_RIGHT, because
glDrawBuffers (which the glDrawBuffer calls) doesn't accept
GL_FRONT!

So this was producing errors, and most code doesn't call
glGetError, but should, so that's why this went unnoticed.

I suspect some video cards handle the error differently, and
that may be why this occurred only with some video drivers.

This was found during usage of the OpenGL Profiler, which
as I understand it didn't work correctly with the old libGL.
It seems to work fine with AppleSGLX though.

So hurray!  It may fix Pymol!



git-svn-id: http://svn.macosforge.org/repository/xquartz/AppleSGLX/trunk@265 9052f27e-8171-453d-b9bb-b0e8541e09c8