Restructure the drawable types to use struct apple_glx_drawable like a base class.
This simplifies the thread locking, and eliminates some code.

This gives us better reference counting for drawables, and locking behavior.

You can now find/retrieve a struct apple_glx_drawable with a GLXDrawable XID, and
the routine is able to return a locked or referenced object.  This prevents
races and should work better.

The changes are pretty significant, but I've done my best to test them with Mesa
demos and the test suite.

I also added a new feature -- env LIBGL_DIAGNOSTIC=1 ./someprog  now enables a
lot more useful constructor and destructor information, as well as information
about surface notify events, and how they were handled.  The way it is implemented
it doesn't continuously check getenv(), so it should be fast enough, and seems to
be fast in practice.



git-svn-id: http://svn.macosforge.org/repository/xquartz/AppleSGLX/trunk@267 9052f27e-8171-453d-b9bb-b0e8541e09c8