Add more diagnostics for surfaces and drawables.
It seems I was wrong about the test code not causing a leak.

The drawables are created and destroyed, but it seems that
the WindowServer grows.  Something appears to be leaking
and I'm not sure what yet.  

I will verify with some printf statements in the X server 
to make sure it's destroying the surface when 
XAppleDRIDestroySurface is called.



git-svn-id: http://svn.macosforge.org/repository/xquartz/AppleSGLX/trunk@278 9052f27e-8171-453d-b9bb-b0e8541e09c8